# src/main/resources/public/api-contract.yml
openapi: 3.0.1
info:
  title: Votação API
  description: |-
    Para obter mais informações sobre o desenvolvimento

    ## Overview

    Api CoopVotes

    Esta api está agnostica de clientes deste que uso protocolos, **HTTP**.

    ## Começando

    Esta documentação ira auxiliar em novas implementações de __clients sides__, dando uma base
    de como manipular os recursos que a API provem.

    ### HTTP Verbs

    Esta Api foi construída a seguir os padrões RESTFull, tentando aderir o mais próximo
    possíveis das convenções uso dos verbos HTTP.


    | Verbo     | Utilização                                      |
    | :-------: | :---------------------------------------------- |
    | `GET`     | Usado para recuperar recursos                   |
    | `POST`    | Usado para criar um recurso                     |
    | `PUT`     | Usado para atualizar um recurso                 |
    | `PATCH`   | Usado para atualizar um parcialmente um recurso |
    | `DELETE`  | Usado para remover um recurso existente         |

    ### HTTP status codes

    Esta Api foi construída seguindo os padrões RESTFull, tentando aderir o mais próximo
    possíveis as suas convenções em seu uso dos códigos de Status HTTP.

    | Status code                    | Utilização                                                                                                                          |
...

    # Header
...
       aqui dispostos.

    ## Requisições
    Headers para enviar requisições.

...

...

...

...

...

...

...

...
servers:
...
paths:
  /staves:
    post:
      summary: Cria uma nova Pauta
      operationId: create
      tags:
        - Stave
      description: Cria um recurso Pauta `Stave`
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InputNewStave'
      responses:
        '201':
          $ref: '#/components/responses/StaveCreateResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      summary: Atualiza uma Pauta
      operationId: update
      tags:
        - Stave
      description: |-
        Uma Pauta pode ser atualizada respeita as seguintes regras:

        * A pauta poderá ser atualizada so e somente só, estiver no estado __CREATE__;
        * O tema `theme`, apenas se tema não estiver cadastrado posteriormente, exceto a Pauta com
        o tema cadastrado anteriormente estive no estado finalizado __SESSION_VOTES_DONE__;
      parameters:
        - in: path
          name: id
          description: identificar único Pauta encontrar a pauta a ser atualizada
          schema:
            type: integer
            format: int64
            example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InputUpdateStave'
      responses:
        '200':
          $ref: '#/components/responses/StaveUpdateResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      summary: Cancela uma Pauta
      operationId: cancel
      tags:
        - Stave
      description: |-
        Uma Pauta pode ser cancelada respeita as seguinte regra:

        * A pauta poderá ser cancelada so e somente só, estiver no estado __CREATE__ ou __VOTING_IN_PROGRESS__;
      parameters:
        - in: path
          name: id
          description: identificar único Pauta para localizar a Pauta a ser cancelada.
          schema:
            type: integer
            format: int64
            example: 1
      responses:
        '204':
          description: Pauta cancelada com sucesso.
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /start-session-vote:
    patch:
      summary: Inicia uma sessão para votação de uma pauta especifica.
      operationId: startSession
      description: Inicia uma sessão de votação. Lembre-se caso o timeToLive não for informado, o valor padrão será utilizado, neste caso o valor é 1 (um) minuto.
      tags:
        - Stave
      parameters:
        - in: path
          name: id
          description: identificar único Pauta para iniciar a sessão.
          schema:
            type: integer
            format: int64
            example: 1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InputStartSession'
      responses:
        '202':
          description: Sessão de votação iniciada com sucesso
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  schemas:
...
    Stave:
      properties:
        id:
          type: integer
          format: int64
          description: Identificador único de uma Pauta
          example: 1001
        theme:
          type: string
          description: Tema da pauta
          example: Pauta que será discutida tal assunto.
        description:
          type: string
          description: Texto descrição da pauta
          example: Descrição da Pauta para o tema tal.
        state:
          type: string
          description: Estado da pauta
          example: create
        createAt:
          type: string
          format: date-time
          description: Data de criação da Pauta
          example: "2021-01-30T08:30:00Z"
        startSessionVoting:
          type: string
          format: date-time
          description: Data do inicio da sessão de votação da Pauta
          example: "2021-01-30T08:30:00Z"
        endSessionVoting:
          type: string
          format: date-time
          description: Data do fim da sessão de votação da Pauta
          example: "2021-01-30T08:30:00Z"
        totalVotes:
          type: integer
          description: total de votos que a Pauta teve na sessão de votação.
          example: 500
        totalVotesYes:
          type: integer
          description: total de votos positivos que a Pauta teve.
          example: 300
        totalVotesNo:
          type: integer
          description: total de votos negativos que a Pauta teve.
          example: 200

    #Inputs
    InputNewStave:
      type: object
      required:
        - theme
        - description
      properties:
        theme:
          type: string
          minLength: 5
          maxLength: 100
          description: Texto simples para que defini o tema da pauta
          example: Pauta para definir o assunto tal.
        description:
          type: string
          minLength: 5
          maxLength: 255
          description: Texto descritivo para melhor explicação referente ao tema.
          example: Pauta foi criada para melhor elaborar tal assunto.
    InputStartSession:
      type: object
      properties:
        timeToLive:
          type: integer
          format: int64
          default: 60
          minimum: 1
          description: Tempo em seguntos da duração de uma sessão.
          example: 10
    InputUpdateStave:
      properties:
        theme:
          type: string
          description: Tema da pauta
          example: Pauta que será discutida tal assunto.
        description:
          type: string
          description: Texto descrição da pauta
          example: Descrição da Pauta para o tema tal.
...
    ErrorBase:
      type: object
      properties:
        statusCode:
          type: integer
          example: 0
          description: Code de status http
        message:
          type: string
          example: Um erro ocorreu.
          description: Mensagem da ocorrencia do erro.
        path:
          type: string
          example: /path
          description: uri da ocorrencia do erro
    Error:
      allOf:
        - type: object
        - $ref: '#/components/schemas/ErrorBase'
        - properties:
            details:
              type: array
              items:
                $ref: '#/components/schemas/ErrorDetails'
    ErrorDetails:
      type: object
      properties:
        field:
          type: string
          example: type.field
          description: nome do campo que contém um erro
        descriptionError:
          type: string
          example: Campo inválido
          description: descreve o erro encontrado no campo.
  headers:
    Version-Api:
      description: Define a versão da api, da qual respondeu a solicitação.
      schema:
        type: string
        default: '1.0.0'
      example: '0.0.1'
  responses:
    StaveCreateResponse:
      description: Retorno de uma solicitação para ação de criar uma Pauta com sucesso.
      headers:
        Version-Api:
          $ref: '#/components/headers/Version-Api'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Stave'
    StaveUpdateResponse:
      description: Retorno de uma solicitação para ação de atualizar uma Pauta com sucesso.
      headers:
        Version-Api:
          $ref: '#/components/headers/Version-Api'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Stave'


    # errors
    BadRequest:
      description: >
        Requisição está invalida, o corpo da mensagem está mal formatada, por exemplo, campos faltando ou
        campos estão inválidos (formato inválido).
          No corpo da resposta encontrar uma lista de detalhes que dará informações pertinentes para corrigir o erro.
        Para maiores informações consulte a [rfc-7231](https://datatracker.ietf.org/doc/html/rfc7231#section-6.5.1).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
      headers:
        Version-Api:
          $ref: '#/components/headers/Version-Api'
    NotFound:
      description: |-
        Recurso não foi encontrado, geralmente a url pode está errada ou recurso pedido não existe.


        Para maiores informações consulte a [rfc-7231](https://httpwg.org/specs/rfc7231.html#status.404).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBase'
    Conflict:
      description: |-
        `Requisição` foi aceita mas, os dados enviados está em conflito com o recurso a ser manipulado
        por exemplo: uma Pauta com um tema já foi cadastrada anteriormente.


        Para maiores informações consulte a [rfc-7231](https://httpwg.org/specs/rfc7231.html#status.409).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBase'
    UnprocessableEntity:
      description: >
        `Requisição` foi aceita mas, não foi possível concluir com sucesso, devido __alguma regra de negocio__
        por exemplo: validações, ou estado de uma Pauta não permite a ação que originou o error.


        Para maiores informações consulte a [rfc-4918](https://datatracker.ietf.org/doc/html/rfc4918#section-11.2).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBase'
    InternalServerError:
      description: >
        Error inesperado em alguma parte do processo, neste caso a api não conseguiu se recuperar
        do error.


        Caso isto ocorra, por favor, notifique o momento do erro, crie uma issue no __repositório__ da aplicação explicando
        os detalhes de como aconteceu o error.


        Para maiores informações consulte a [rfc-7231](https://datatracker.ietf.org/doc/html/rfc7231#section-6.6.1).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBase'